search/build: fix plane batch predicates + exhaustive, purge misaligned typed access
- search_batch plane phase: mirror the solo plane path for predicate
  queries (resolve pred columns, summary-prune candidates, selectivity
  -> FallbackFiltered, adaptive W). Previously s.pred_col_indices stayed
  empty and the sweep harvest's eval_all_predicates subscripts it ->
  SIGSEGV on plane-attached trees (batch + predicates).
- plane routing honors exhaustive (n_probe=UINT32_MAX): fraction budget
  forced to 1.0 and survivor pre-prune disabled, so exact-rerank truth
  is reachable beyond the plane's top-fraction leaves.
- row_ids are NOT 8-aligned in scalar-family leaves: load_rowid/
  store_rowid (memcpy) replace every typed RowId access; staged-record
  fp16 fields are byte-packed at arbitrary offsets -> StagedRow::fp16_vec
  is now const uint8_t* and the build memcpy-appends.
- HLL registers: one byte per register (the 6-bit packing read/wrote one
  byte past the 3072B buffer on the last register; never serialized, so
  no format impact).
- local_scalar encode_one saturates NaN/inf before the int cast.
- LeafExtentCache: alias-run entries live under multiple map keys ->
  dedupe in invalidate_all() and ~Shard() (second key touched the freed
  Entry's flags = UAF).
- ctpl TLS worker tags: mark as intentional leak for LSAN.
- test-side leaks: make_float_view buffer pool, block-cache drain
  off-by-one.

meson test: 24/24 (debug ASAN+UBSAN w/ MALLOC_PERTURB_; quantizer_families
and tree_insert_delete need --timeout 900 under -O0+ASAN) and 24/24 release.

+++ src/tree/cluster_stage.hpp
@@ -72,7 +72,9 @@ struct StagedRow {
     uint16_t flags = 0;
     const uint8_t* code = nullptr;      // code bytes or null
     uint32_t code_len = 0;
-    const float16_t* fp16_vec = nullptr;  // dim floats or null
+    const uint8_t* fp16_vec = nullptr;    // dim halves, BYTE-PACKED at an
+                                           // arbitrary (possibly odd) record
+                                           // offset — copy with memcpy only
     uint32_t fp16_vec_bytes = 0;
     bool has_ip_bias = false;
     float16_t ip_bias{1.0f};
